backend-fastapi/requirements.txt
fastapi
uvicorn
chromadb
google-generativeai
pandas
python-dev
